terminusaJet/requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.7.0
APScheduler==3.10.4
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
eventlet==0.36.1
Faker==25.2.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
magic-filter==1.0.12
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
multidict==6.0.5
mysql-connector==2.2.9
packaging==24.0
pluggy==1.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.2
uvicorn==0.30.1
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
